Refactored the sendTransaction function to be on the table.js page w/ the other CRUD functions
